requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pocketbase==0.12.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
psutil==6.1.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
